+++ support/iam/src/basic/dto.rs
@@ -3,7 +3,7 @@ pub mod iam_app_dto;
 pub mod iam_attr_dto;
 pub mod iam_cert_conf_dto;
 pub mod iam_cert_dto;
-pub mod iam_filter_dto;
+pub mod iam_filer_dto;
 pub mod iam_res_dto;
 pub mod iam_role_dto;
 pub mod iam_set_dto;
