+++ support/iam/tests/test_ca_app.rs
@@ -5,7 +5,7 @@ use tardis::log::info;
 
 use bios_basic::rbum::serv::rbum_item_serv::RbumItemCrudOperation;
 use bios_iam::basic::dto::iam_app_dto::IamAppAggModifyReq;
-use bios_iam::basic::dto::iam_filter_dto::IamAppFilterReq;
+use bios_iam::basic::dto::iam_filer_dto::IamAppFilterReq;
 use bios_iam::basic::serv::iam_app_serv::IamAppServ;
 use bios_iam::iam_constants;
